change few methods to work with frontend

+++ src/main/java/com/glinka/wcn/service/impl/GroupServiceImpl.java
@@ -1,10 +1,12 @@
 package com.glinka.wcn.service.impl;
 
+import com.glinka.wcn.commons.NotAuthorizedException;
 import com.glinka.wcn.commons.ResourceNotFoundException;
 import com.glinka.wcn.model.dao.Group;
 import com.glinka.wcn.model.dao.Journal;
 import com.glinka.wcn.model.dao.User;
 import com.glinka.wcn.model.dto.GroupDto;
+import com.glinka.wcn.model.dto.GroupNameDto;
 import com.glinka.wcn.model.dto.ScientificJournalDto;
 import com.glinka.wcn.repository.GroupRepository;
 import com.glinka.wcn.repository.ScientificJournalRepository;
@@ -15,6 +17,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -23,14 +27,16 @@ import java.util.stream.Collectors;
 public class GroupServiceImpl implements GroupService {
 
     private final Mapper<GroupDto, Group> groupMapper;
+    private final Mapper<GroupNameDto, Group> groupNameMapper;
     private final Mapper<ScientificJournalDto, Journal> scientificJournalMapper;
     private final GroupRepository groupRepository;
     private final ScientificJournalRepository scientificJournalRepository;
     private final UserRepository userRepository;
 
     @Autowired
-    public GroupServiceImpl(Mapper<GroupDto, Group> groupMapper, Mapper<ScientificJournalDto, Journal> scientificJournalMapper, GroupRepository groupRepository, ScientificJournalRepository scientificJournalRepository, UserRepository userRepository) {
+    public GroupServiceImpl(Mapper<GroupDto, Group> groupMapper, Mapper<GroupNameDto, Group> groupNameMapper, Mapper<ScientificJournalDto, Journal> scientificJournalMapper, GroupRepository groupRepository, ScientificJournalRepository scientificJournalRepository, UserRepository userRepository) {
         this.groupMapper = groupMapper;
+        this.groupNameMapper = groupNameMapper;
         this.scientificJournalMapper = scientificJournalMapper;
         this.groupRepository = groupRepository;
         this.scientificJournalRepository = scientificJournalRepository;
@@ -56,8 +62,12 @@ public class GroupServiceImpl implements GroupService {
     }
 
     @Override
-    public GroupDto save(GroupDto groupDto) {
-        return groupMapper.mapToDto(groupRepository.saveAndFlush(groupMapper.mapToDao(groupDto)));
+    public GroupDto save(GroupNameDto groupDto, String email) {
+        User owner = userRepository.findUserByEmail(email);
+        Group group = groupNameMapper.mapToDao(groupDto);
+        group.setOwner(owner);
+        group.setUsers(Collections.singletonList(owner));
+        return groupMapper.mapToDto(groupRepository.saveAndFlush(group));
     }
 
     @Override
@@ -78,14 +88,13 @@ public class GroupServiceImpl implements GroupService {
     }
 
     @Override
-    public GroupDto addUser(Long userId, Long groupId) throws ResourceNotFoundException  {
+    public GroupDto addUser(String owner, String email, Long groupId) throws ResourceNotFoundException, NotAuthorizedException {
         Group group = groupRepository.findById(groupId).orElseThrow(
                 () -> new ResourceNotFoundException("Group with id: " + groupId + " not found")
         );
+        isOwner(owner, group);
         List<User> userList = group.getUsers();
-        User newUser = userRepository.findById(userId).orElseThrow(
-                () -> new ResourceNotFoundException("User with id: " + userId + " not found.")
-        );
+        User newUser = userRepository.findUserByEmail(email);
         if (userList.contains(newUser)){
             return groupMapper.mapToDto(group);
         }
@@ -95,10 +104,11 @@ public class GroupServiceImpl implements GroupService {
     }
 
     @Override
-    public void removeUser(Long userId, Long groupId) throws ResourceNotFoundException {
+    public void removeUser(String owner, Long userId, Long groupId) throws ResourceNotFoundException, NotAuthorizedException {
         Group group = groupRepository.findById(groupId).orElseThrow(
                 () -> new ResourceNotFoundException("Group with id: " + groupId + " not found")
         );
+        isOwner(owner, group);
         List<User> userList = group.getUsers();
         userList.remove(userRepository.findById(userId).orElseThrow(
                 () -> new ResourceNotFoundException("User with id: " + userId + " not found.")
@@ -128,11 +138,9 @@ public class GroupServiceImpl implements GroupService {
     }
 
     @Override
-    public List<GroupDto> findAllByUser(Long userId) throws ResourceNotFoundException {
+    public List<GroupDto> findAllByUser(String email) throws ResourceNotFoundException {
         List<Group> groups = groupRepository.findAll();
-        User user = userRepository.findById(userId).orElseThrow(
-                () -> new ResourceNotFoundException("User with id: " + userId + " not found.")
-        );
+        User user = userRepository.findUserByEmail(email);
         List<Group> groupDaosByUser = new ArrayList<>();
         for(Group group : groups){
             if(group.getUsers().contains(user)){
@@ -143,8 +151,29 @@ public class GroupServiceImpl implements GroupService {
     }
 
     @Override
-    public void deleteGroup(Long id) throws ResourceNotFoundException {
-        findById(id);
-        groupRepository.deleteById(id);
+    public List<GroupNameDto> findAllGroupsNameByUser(String email) throws ResourceNotFoundException {
+        List<Group> groups = groupRepository.findAll();
+        User user = userRepository.findUserByEmail(email);
+        List<Group> groupDaosByUser = new ArrayList<>();
+        for(Group group : groups){
+            if(group.getOwner().equals(user)){
+                groupDaosByUser.add(group);
+            }
+        }
+        return groupDaosByUser.stream().map(groupNameMapper::mapToDto).collect(Collectors.toList());
+    }
+
+    @Override
+    public void deleteGroup(String owner, Long id) throws ResourceNotFoundException, NotAuthorizedException {
+        if (findById(id).getOwnerDto().getEmail().equals(owner)){
+            groupRepository.deleteById(id);
+        } else throw new NotAuthorizedException("This is not our group");
+    }
+
+    private void isOwner(String owner, Group group) throws NotAuthorizedException {
+        User userInGroup = group.getUsers().stream().filter(user -> owner.equals(user.getEmail())).findAny().orElse(null);
+        if (userInGroup == null || !group.getOwner().getEmail().equals(owner) || !owner.equals(userInGroup.getEmail())){
+            throw new NotAuthorizedException("User with email: " + owner + " can't add new user to group with id: " + group.getGroupId());
+        }
     }
 }
